requirements.txt
attrs==21.2.0
certifi==2021.10.8
charset-normalizer==2.0.7
codecov==2.1.12
coverage==6.1.1
flake8==4.0.1
idna==3.3
iniconfig==1.1.1
mccabe==0.6.1
packaging==21.2
pluggy==1.0.0
psycopg2-binary==2.9.2
py==1.11.0
pycodestyle==2.8.0
pyflakes==2.4.0
pymongo==3.12.1
pyparsing==2.4.7
pytest==6.2.5
pytest-cov==3.0.0
requests==2.26.0
toml==0.10.2
tomli==1.2.2
urllib3==1.26.7
